+++ tests/python/sdk/util.py
@@ -4,8 +4,11 @@
 
 import textwrap
 from pathlib import Path
-from typing import Tuple
+from typing import Container, Tuple
+from urllib.parse import urlparse
 
+import pytest
+from cvat_sdk.api_client.rest import RESTClientObject
 from cvat_sdk.core.helpers import TqdmProgressReporter
 from tqdm import tqdm
 
@@ -82,3 +85,17 @@ def generate_coco_anno(image_path: str, image_width: int, image_height: int) -> 
             "image_width": image_width,
         }
     )
+
+
+def restrict_api_requests(
+    monkeypatch: pytest.MonkeyPatch, allow_paths: Container[str] = ()
+) -> None:
+    original_request = RESTClientObject.request
+
+    def restricted_request(self, method, url, *args, **kwargs):
+        parsed_url = urlparse(url)
+        if parsed_url.path in allow_paths:
+            return original_request(self, method, url, *args, **kwargs)
+        raise RuntimeError("Disallowed!")
+
+    monkeypatch.setattr(RESTClientObject, "request", restricted_request)